requirements.txt
Cartopy==0.17.0
cftime==1.1.0
Click==7.0
cycler==0.10.0
ecmwf-api-client==1.5.4
Flask==1.1.1
geos==0.2.2
iris==1.0.7
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.3
netCDF4==1.5.3
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
pyparsing==2.4.6
pyshp==2.1.0
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
Werkzeug==1.0.0
xarray==0.15.0
